DeliveryLocator/urls.py
"""
URL configuration for DeliveryLocator project.

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/4.2/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
from django.conf import settings
from django.conf.urls.static import static
from django.contrib import admin
from django.urls import include, path
from drf_yasg import openapi
from drf_yasg.views import get_schema_view
from rest_framework import permissions

schema_view = get_schema_view(
    openapi.Info(
        title="Customer Delivery Location API",
        default_version="v1",
        description="This API endpoint, avaliable at 'localhost:8000/api/v1/delivery-locations', is designed to retrieve a list of all delivery locations stored in the system. It provides access to comprehensive information about each delivery location, including its name, geographic coordinates, and additional details.",
        terms_of_service="https://https://github.com/ishanshre/GeoDjangoCLDM",
        contact=openapi.Contact(email="ishan.shrestha356@gmail.com"),
    ),
    public=True,
    permission_classes=(permissions.AllowAny,),
)

urlpatterns = [
    path("admin/", admin.site.urls),
    path("", include("location.urls")),
    path("api/v1/", include("location.api.urls")),
    path("user/", include("account.urls")),
    path("__debug__/", include("debug_toolbar.urls")),
    path(
        "api/docs/swagger<format>/",
        schema_view.without_ui(cache_timeout=0),
        name="schema-json",
    ),
    path(
        "api/docs/swagger/",
        schema_view.with_ui("swagger", cache_timeout=0),
        name="schema-swagger-ui",
    ),
    path(
        "api/docs/redoc/",
        schema_view.with_ui("redoc", cache_timeout=0),
        name="schema-redoc",
    ),
]

if settings.DEBUG:
    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
