requirements.txt
asgiref==3.7.2
Babel==2.12.1
crispy-bootstrap5==0.7
Django==4.2.4
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-debug-toolbar==4.2.0
django-filter==23.2
django-leaflet==0.29.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
GDAL==3.7.1.1
install==1.3.5
Markdown==3.4.4
phonenumbers==8.13.18
Pillow==10.0.0
psycopg2==2.9.7
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
